requirements.txt
altgraph==0.17.4
attrs==24.3.0
auto-py-to-exe==2.45.1
beautifulsoup4==4.12.3
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cx_Freeze==7.2.8
cx_Logging==3.2.1
Eel==0.18.1
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
lief==0.16.2
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.28.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
cassandra-driver
fastapi
uvicorn
pdfplumber
uuid
langchain
langchain-community
langchain-experimental
huggingface-hub
torch
transformers
sentence-transformers
python-multipart
cassio